Refactor semantic tokens decoding
