+++ packages/langium/src/grammar/langium-grammar-semantic-token-provider.ts
@@ -4,8 +4,9 @@
  * terms of the MIT License, which is available in the project root.
  ******************************************************************************/
 
-import { SemanticTokenTypes } from 'vscode-languageserver';
-import { AbstractSemanticTokenProvider, SemanticTokenAcceptor } from '../lsp/semantic-token-provider';
+import { SemanticTokens, SemanticTokenTypes } from 'vscode-languageserver';
+import { LangiumDocument } from '..';
+import { AbstractSemanticTokenProvider, AllSemanticTokenTypes, SemanticTokenAcceptor } from '../lsp/semantic-token-provider';
 import { AstNode } from '../syntax-tree';
 import { isAction, isAssignment, isAtomType, isParameter, isParameterReference, isReturnType } from './generated/ast';
 
@@ -55,4 +56,47 @@ export class LangiumGrammarSemanticTokenProvider extends AbstractSemanticTokenPr
         }
     }
 
+}
+
+export interface DecodedSemanticToken {
+    line: number;
+    character: number;
+    length: number;
+    tokenType: SemanticTokenTypes;
+    text: string;
+    tokenModifiers: number;
+}
+
+export class SemanticTokensDecoder {
+    static decode<T extends AstNode=AstNode>(tokens: SemanticTokens, document: LangiumDocument<T>): DecodedSemanticToken[] {
+        const typeMap = new Map<number, SemanticTokenTypes>();
+        Object.entries(AllSemanticTokenTypes).forEach(([type, index]) => typeMap.set(index, type as SemanticTokenTypes));
+        let line = 0;
+        let character = 0;
+        return this.sliceIntoChunks(tokens.data, 5).map(t => {
+            line += t[0];
+            if (t[0] !== 0) {
+                character = 0;
+            }
+            character += t[1];
+            const length = t[2];
+            return {
+                line,
+                character,
+                length,
+                tokenType: typeMap.get(t[3])!,
+                tokenModifiers: t[4],
+                text: document.textDocument.getText({ start: { line, character }, end: { line, character: character + length } })
+            } as DecodedSemanticToken;
+        });
+    }
+
+    private static sliceIntoChunks<T>(arr: T[], chunkSize: number) {
+        const res = [];
+        for (let i = 0; i < arr.length; i += chunkSize) {
+            const chunk = arr.slice(i, i + chunkSize);
+            res.push(chunk);
+        }
+        return res;
+    }
 }
